Write Stack
Write and test a new class called Stack that contains most of the logic
for the gem. In the process, add a few custom error classes and a few
test middleware and definition classes. Also, be sure to require
'middlegem/stack' and all the custom error classes in 'middlegem.rb'.